Message Service
MessageService manages channel message communication in Angular using WebSockets and reactive signals. It keeps a list of messages in the messages signal and a local queue messageQueue for sent messages. It allows initializing a connection to a channel with init(channelId), sending messages with send(text), and listens to external notifications of type message_sent to automatically add messages to the signal. The WebSocket callbacks (onOpen, onClose, onMessage, onError) handle connection events, and NgZone ensures that signal updates occur within the Angular context, all of this only in the browser.